+++ src/settings.ts
@@ -261,6 +261,19 @@ export default class HeatmapTrackerSettingsTab extends PluginSettingTab {
         }));
   }
 
+  private displayChristmasSettings() {
+    const { containerEl } = this;
+    new Setting(containerEl)
+      .setName("🎄 Christmas mood")
+      .setDesc("Enable snowfall effect and Santa Claus hat on the heatmap tracker to have Christmas vibes in your Obsidian. Let it snow, let it snow, let it snow!")
+      .addToggle(toggle => toggle
+        .setValue(this.plugin.settings.enableChristmasMood)
+        .onChange(async (value) => {
+          this.plugin.settings.enableChristmasMood = value;
+          await this.plugin.saveSettings();
+        }));
+  }
+
   display() {
     const { containerEl } = this;
 
@@ -271,6 +284,8 @@ export default class HeatmapTrackerSettingsTab extends PluginSettingTab {
     this.displayWeekStartDaySettings();
 
     this.displaySeparateMonthsSettings();
+
+    this.displayChristmasSettings();
 
     this.displayPaletteSettings();
   }