+++ src/views/HeatmapTrackerView/HeatmapTrackerView.tsx
@@ -57,6 +57,12 @@ export function HeatmapTrackerView() {
  * IN PROGRESS
  */
 function HeatmapWeekNumbers() {
+  const {
+    trackerData,
+  } = useHeatmapContext();
+
+  const separateMonths = trackerData.separateMonths;
+
   return (
     <div className="heatmap-tracker-week-numbers">
       {Array.from({ length: 53 }, (_, i) => (